✏️ Rename: `members.ts` to `MEMBERS.ts`
