+++ README.md
@@ -17,7 +17,7 @@
 - [仕様書(雑)](./docs/SPECIFICATION.md)
 
 * [ニュース記事の追加方法](./docs/ADDING_NEWS.md)
-* [メンバーのリスト判定への追加](./src/lib/data/members.ts)
+* [メンバーのリスト判定への追加](./src/lib/data/MEMBERS.ts)
 
 ---
